--- strongbox-commons/src/main/java/org/carlspring/strongbox/data/config/CommonConfig.java
@@ -1,20 +1,0 @@
-package org.carlspring.strongbox.data.config;
-
-import org.springframework.context.annotation.ComponentScan;
-import org.springframework.context.annotation.Configuration;
-
-@Configuration
-@ComponentScan({
-        "org.carlspring.strongbox.configuration",
-        "org.carlspring.strongbox.io",
-        "org.carlspring.strongbox.net",
-        "org.carlspring.strongbox.resource",
-        "org.carlspring.strongbox.rest",
-        "org.carlspring.strongbox.storage.repository",
-        "org.carlspring.strongbox.url",
-        "org.carlspring.strongbox.util",
-        "org.carlspring.strongbox.xml"
-})
-public class CommonConfig
-{
-}
