--- strongbox-data-service/src/main/java/org/carlspring/strongbox/data/config/DataServicePropertiesConfig.java
@@ -1,18 +1,0 @@
-package org.carlspring.strongbox.data.config;
-
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.context.annotation.PropertySource;
-import org.springframework.context.support.PropertySourcesPlaceholderConfigurer;
-
-@Configuration
-@PropertySource(value = { "classpath:META-INF/strongbox-db.properties" })
-public class DataServicePropertiesConfig
-{
-
-    @Bean
-    public static PropertySourcesPlaceholderConfigurer propertySourcesPlaceholderConfigurer()
-    {
-        return new PropertySourcesPlaceholderConfigurer();
-    }
-}
