--- strongbox-data-service/src/main/java/org/carlspring/strongbox/data/repository/UserRepository.java
@@ -1,20 +1,0 @@
-package org.carlspring.strongbox.data.repository;
-
-import org.carlspring.strongbox.data.domain.User;
-
-import org.springframework.transaction.annotation.Transactional;
-
-/**
- * Spring Data CRUD repository for {@link User}.
- *
- * @author Alex Oreshkevich
- */
-@Transactional
-public interface UserRepository
-        extends OrientRepository<User>
-{
-
-    // select * from User where username = 'admin'
-    User findByUsername(String userName);
-}
-
