+++ strongbox-data-service/src/main/java/org/carlspring/strongbox/users/data/repository/OrientRepository.java
@@ -1,0 +1,16 @@
+package org.carlspring.strongbox.users.data.repository;
+
+import org.springframework.data.orient.object.repository.OrientObjectRepository;
+import org.springframework.data.repository.NoRepositoryBean;
+
+/**
+ * Marker interface for defining CRUD API under any Object based on OrientDB storage.
+ *
+ * @author Alex Oreshkevich
+ */
+@NoRepositoryBean
+public interface OrientRepository<T>
+        extends OrientObjectRepository<T>
+{
+
+}
