--- strongbox-data-service/src/test/java/org/carlspring/strongbox/data/DataServiceTestContext.java
@@ -1,27 +1,0 @@
-package org.carlspring.strongbox.data;
-
-import org.carlspring.strongbox.data.config.DataServiceConfig;
-
-import java.lang.annotation.ElementType;
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-import java.lang.annotation.Target;
-
-import org.springframework.test.annotation.Rollback;
-import org.springframework.test.context.ContextConfiguration;
-import org.springframework.transaction.annotation.Transactional;
-
-/**
- * Meta annotation for generalised test execution context management.
- *
- * @author Alex Oreshkevich
- */
-@Target(ElementType.TYPE)
-@Retention(RetentionPolicy.RUNTIME)
-@ContextConfiguration(classes = { DataServiceConfig.class })
-@Transactional
-@Rollback // change to false if you wanna inspect database content manually
-public @interface DataServiceTestContext
-{
-
-}
