+++ strongbox-security-api/src/test/java/org/carlspring/strongbox/security/certificates/KeyStoresTest.java
@@ -1,6 +1,6 @@
 package org.carlspring.strongbox.security.certificates;
 
-import org.carlspring.strongbox.data.config.StrongboxSecurityConfig;
+import org.carlspring.strongbox.users.data.config.StrongboxSecurityConfig;
 import org.carlspring.strongbox.net.ConnectionChecker;
 import org.carlspring.strongbox.testing.AssignedPorts;
 import org.junit.Before;
