+++ strongbox-storage/strongbox-storage-api/src/test/java/org/carlspring/strongbox/configuration/ConfigurationManagerTest.java
@@ -1,7 +1,7 @@
 package org.carlspring.strongbox.configuration;
 
-import org.carlspring.strongbox.data.config.CommonConfig;
-import org.carlspring.strongbox.data.config.StorageApiConfig;
+import org.carlspring.strongbox.users.data.config.CommonConfig;
+import org.carlspring.strongbox.users.data.config.StorageApiConfig;
 import org.carlspring.strongbox.services.ArtifactResolutionService;
 import org.carlspring.strongbox.storage.Storage;
 import org.carlspring.strongbox.storage.repository.Repository;
