+++ strongbox-storage/strongbox-storage-api/src/test/java/org/carlspring/strongbox/storage/resolvers/FSLocationResolverTest.java
@@ -1,7 +1,7 @@
 package org.carlspring.strongbox.storage.resolvers;
 
-import org.carlspring.strongbox.data.config.CommonConfig;
-import org.carlspring.strongbox.data.config.StorageApiConfig;
+import org.carlspring.strongbox.users.data.config.CommonConfig;
+import org.carlspring.strongbox.users.data.config.StorageApiConfig;
 import org.carlspring.strongbox.configuration.ConfigurationManager;
 import org.carlspring.strongbox.resource.ConfigurationResourceResolver;
 import org.carlspring.strongbox.storage.repository.Repository;
