+++ strongbox-storage/strongbox-storage-api/src/test/java/org/carlspring/strongbox/storage/resolvers/GroupLocationResolverTest.java
@@ -2,8 +2,8 @@ package org.carlspring.strongbox.storage.resolvers;
 
 import com.carmatechnologies.commons.testing.logging.ExpectedLogs;
 import com.carmatechnologies.commons.testing.logging.api.LogLevel;
-import org.carlspring.strongbox.data.config.CommonConfig;
-import org.carlspring.strongbox.data.config.StorageApiConfig;
+import org.carlspring.strongbox.users.data.config.CommonConfig;
+import org.carlspring.strongbox.users.data.config.StorageApiConfig;
 import org.carlspring.strongbox.configuration.ConfigurationManager;
 import org.carlspring.strongbox.resource.ConfigurationResourceResolver;
 import org.carlspring.strongbox.resource.ResourceCloser;
