+++ strongbox-storage/strongbox-storage-api/src/test/java/org/carlspring/strongbox/storage/validation/version/VersionValidatorServiceTest.java
@@ -1,7 +1,7 @@
 package org.carlspring.strongbox.storage.validation.version;
 
-import org.carlspring.strongbox.data.config.CommonConfig;
-import org.carlspring.strongbox.data.config.StorageApiConfig;
+import org.carlspring.strongbox.users.data.config.CommonConfig;
+import org.carlspring.strongbox.users.data.config.StorageApiConfig;
 import org.carlspring.strongbox.services.VersionValidatorService;
 import org.junit.Test;
 import org.junit.runner.RunWith;
