--- strongbox-testing/strongbox-testing-core/src/main/java/org/carlspring/strongbox/data/config/TestingCoreConfig.java
@@ -1,17 +1,0 @@
-package org.carlspring.strongbox.data.config;
-
-import org.carlspring.strongbox.testing.AssignedPorts;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-
-@Configuration
-public class TestingCoreConfig
-{
-
-    @Bean(name = "assignedPorts")
-    AssignedPorts assignedPorts()
-    {
-        return new AssignedPorts();
-    }
-
-}
