+++ strongbox-user-management/strongbox-user-management-api/src/main/java/org/carlspring/strongbox/users/repository/UserRepository.java
@@ -1,0 +1,21 @@
+package org.carlspring.strongbox.users.repository;
+
+import org.carlspring.strongbox.users.data.repository.OrientRepository;
+import org.carlspring.strongbox.users.domain.User;
+
+import org.springframework.transaction.annotation.Transactional;
+
+/**
+ * Spring Data CRUD repository for {@link User}.
+ *
+ * @author Alex Oreshkevich
+ */
+@Transactional
+public interface UserRepository
+        extends OrientRepository<User>
+{
+
+    // select * from User where username = 'admin'
+    User findByUsername(String userName);
+}
+
