+++ strongbox-user-management/strongbox-user-management-api/src/main/java/org/carlspring/strongbox/users/service/UserService.java
@@ -1,0 +1,22 @@
+package org.carlspring.strongbox.users.service;
+
+import org.carlspring.strongbox.users.domain.User;
+import org.carlspring.strongbox.users.data.service.CrudService;
+
+import java.util.Optional;
+
+import org.springframework.transaction.annotation.Transactional;
+
+/**
+ * CRUD service for managing {@link User} entities.
+ *
+ * @author Alex Oreshkevich
+ */
+@Transactional
+public interface UserService
+        extends CrudService<User, String>
+{
+
+    @Transactional
+    Optional<User> findByUserName(final String username);
+}
