+++ strongbox-user-management/strongbox-user-management-api/src/test/java/org/carlspring/strongbox/users/UserServiceTest.java
@@ -1,0 +1,19 @@
+package org.carlspring.strongbox.users;
+
+import org.junit.Test;
+import org.junit.runner.RunWith;
+import org.springframework.test.context.junit4.SpringJUnit4ClassRunner;
+
+/**
+ * @author Alex Oreshkevich
+ */
+@UserServiceTestContext
+@RunWith(SpringJUnit4ClassRunner.class)
+public class UserServiceTest
+{
+
+    @Test
+    public void test(){
+
+    }
+}
