+++ strongbox-web-core/src/main/java/org/carlspring/strongbox/config/SecurityConfig.java
@@ -1,19 +1,18 @@
 package org.carlspring.strongbox.config;
 
 import org.carlspring.strongbox.security.authentication.UnauthorizedEntryPoint;
-import org.carlspring.strongbox.security.roles.Roles;
-import org.carlspring.strongbox.security.user.StrongboxUserDetailService;
+import org.carlspring.strongbox.users.domain.Roles;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
+import org.springframework.security.authentication.AuthenticationProvider;
 import org.springframework.security.config.annotation.authentication.builders.AuthenticationManagerBuilder;
 import org.springframework.security.config.annotation.method.configuration.EnableGlobalMethodSecurity;
 import org.springframework.security.config.annotation.web.builders.HttpSecurity;
 import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
 import org.springframework.security.config.annotation.web.configuration.WebSecurityConfigurerAdapter;
 import org.springframework.security.config.http.SessionCreationPolicy;
-import org.springframework.security.core.userdetails.UserDetailsService;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.security.crypto.password.StandardPasswordEncoder;
 import org.springframework.security.web.AuthenticationEntryPoint;
@@ -29,6 +28,9 @@ import org.springframework.security.web.AuthenticationEntryPoint;
 public class SecurityConfig
         extends WebSecurityConfigurerAdapter
 {
+
+    @Autowired
+    private AuthenticationProvider authenticationProvider;
 
     @Override
     protected void configure(HttpSecurity http)
@@ -54,17 +56,12 @@ public class SecurityConfig
     public void configure(AuthenticationManagerBuilder auth)
             throws Exception
     {
+        // auth.authenticationProvider(authenticationProvider);
         auth
                 .inMemoryAuthentication()
                 .withUser("maven").password("password").roles(Roles.all())
                 .and()
                 .withUser("admin").password("password").roles(Roles.all());
-    }
-
-    @Override
-    public UserDetailsService userDetailsService()
-    {
-        return new StrongboxUserDetailService();
     }
 
     @Bean
