+++ strongbox-web-core/src/main/java/org/carlspring/strongbox/security/user/StrongboxUserDetailService.java
@@ -1,12 +1,11 @@
 package org.carlspring.strongbox.security.user;
 
-import org.carlspring.strongbox.data.domain.User;
-import org.carlspring.strongbox.data.service.UserService;
+import org.carlspring.strongbox.users.domain.User;
+import org.carlspring.strongbox.users.service.UserService;
 
-import javax.annotation.PostConstruct;
-import java.util.Collections;
 import java.util.Optional;
 
+import com.orientechnologies.orient.object.db.OObjectDatabaseTx;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -14,9 +13,11 @@ import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.security.core.userdetails.UserDetailsService;
 import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.security.crypto.password.PasswordEncoder;
+import org.springframework.stereotype.Component;
 import org.springframework.transaction.annotation.Transactional;
 
 @Transactional
+@Component
 public class StrongboxUserDetailService
         implements UserDetailsService
 {
@@ -29,25 +30,8 @@ public class StrongboxUserDetailService
     @Autowired
     private UserService userService;
 
-    @PostConstruct
-    void setup()
-    {
-
-        final String admin = "admin";
-
-        // add a user for testing
-        // Usually OrientDB should be setup separately(i.e.remotely) and
-        // users should be added either via REST endpoint or via OrientDB Studio/Console
-        if (!userService.findByUserName(admin).isPresent())
-        {
-            User user = new User();
-            user.setUsername(admin);
-            user.setPassword(passwordEncoder.encode("password"));
-            user.setRoles(Collections.singletonList("ROLE_ADMIN"));
-            user.setEnabled(true);
-            userService.save(user);
-        }
-    }
+    @Autowired
+    OObjectDatabaseTx databaseTx;
 
     @Override
     @Transactional
@@ -59,18 +43,12 @@ public class StrongboxUserDetailService
         Optional<User> optionalUser = userService.findByUserName(username);
         optionalUser.orElseThrow(() -> new UsernameNotFoundException("Cannot find user with that username"));
 
-        User user = optionalUser.get();
+        User user = databaseTx.detach(optionalUser.get());
 
         logger.info("user roles: {}", user.getRoles());
 
         // extract (detach) user in current transaction
-        SpringSecurityUser springUser = new SpringSecurityUser();
-        springUser.setEnabled(user.isEnabled());
-        springUser.setPassword(user.getPassword());
-        springUser.setRoles(user.getRoles());
-        springUser.setSalt(user.getSalt());
-        springUser.setUsername(user.getUsername());
-
+        SpringSecurityUser springUser = new SpringSecurityUser(user);
         return springUser;
     }
 }
